refactor(database): Replace monolithic index.ts with modular architecture
Reduced database/index.ts from 881 lines to ~120 lines by extracting:
- DatabaseManager: handles initialization and lifecycle
- MigrationRunner: handles schema migrations and seeding
- DatabaseWriter: handles file persistence with retry

Benefits:
✅ Separation of concerns (each module has single responsibility)
✅ Error handling with retry logic throughout
✅ Easier to test (modular components)
✅ Easier to maintain (smaller, focused files)
✅ Backwards compatible (existing API preserved)
✅ Build successful - all imports working

Old monolithic file backed up as index.old.ts

Next: Write tests for new modules

Phase 0.2 in progress...
